endo_backend/requirements.txt
numpy
pillow
keras
tensorflow
pandas
fastapi
uvicorn
python-multipart
boto3
pymongo
python-jose
passlib[bcrypt]
python-dotenv
torch
torchvision
ultralytics